Import aund.
aund was originally intended to be a complete AUN server for Unix, but seems
to have decayed into being a boot server for NetBSD/arm26.  I wrote it in
1998, and have released a couple of snapshots of it since then.  It seems like
a perfect job for othersrc.

AUN (Acorn Universal Networking) is basically a way of running Econet (Acorn's
ancient proprietary LAN, analagous to LocalTalk) over IP.  aund implements the
Econet fileserver protocol and the IP encapsulation.  The code is a mess, but
it's just about good enough for the job at hand.