BeebEm support: I have a theory about those strange packets.
They contain a destination port number of zero. I've just been
looking at
http://www.chiark.greenend.org.uk/~theom/riscos/docs/Acorn-Unix-Econet-Device-Driver-and-Network-Device.pdf
which suggests that that might be a machine type query, so I've
arranged to answer it as if it is. When I get my own network
utilities up and running, I'll be able to put that theory to the
test properly.