BeebEm support: new approach to ingress filtering.
BeebEm allocates a different UDP port for sending from the one it
uses for receiving, so we can't ingress-filter based on the fixed
ports in econet.cfg, but we can at least check that within a single
login session the sending port doesn't _change_, which would
represent a hijack of the session. We do this by including the IP
source address in the data passed out of beebem.c to the rest of the
code, which will then use it to index into its table of clients - so
somebody trying to hijack a connection using a different UDP port
but claiming to be the same Econet station (and, presumably, being
prepared not to care about receiving replies) will find
fs_find_client has never heard of them.