Robustness: complete revamp of directory handling.
The aund config file now explicitly specifies a _root_ directory for
the whole directory hierarchy being served. aund changes directory
to that as the first thing it does, and then deals with all further
pathnames as relative components on top of that - and also carefully
transforms pathnames received from the server to ensure they can
contain no ".." to break out of that area. fs_nametrans.c has been
substantially rewritten, to support this model and also to simplify
processing in the face of confusing things like "&.^".