Bug fix: robustness in FTSENT handling.
Depending on whether we fts_opened a single file or were going
through a directory, sometimes fts_accpath is the full pathname of
the file, but sometimes it's a prefix to be stuck on the front of
fts_name. Dispense with the use of dirname(3) and instead manually
find the place to apply the suffix fts_name. This fixes the longtxt
examine mode, which was returning the wrong metadata due to this.