Configurability: separate *OPT 4 per user.
The 'opt4' directive in the config file still exists, but I've
relegated it to the status of a default to be used in the absence of
a password file, or if the password file doesn't contain a *OPT 4
entry for that user. (For ease of migration from existing password
files, I've made the trailing OPT 4 field on each line optional.)

This change does not include the ability to actually set your boot
option; it just means that if it's in the password file, it'll be
used.