Robustness: clip filenames at 10 characters.
A new transformation in fs_match_path means that the clipping should
apply to every element in an entire pathname. I've also abstracted
out the informal hidden-file tests in several other pieces of the
code and turned them into calls to an 'is this file hidden?'
function in fs_nametrans.c, which now returns true for >10-character
file names as well as for dotfiles. So aund should now provide a
consistent view of the file system which includes only those file
names it can sensibly display, and the only oddity will be that if
you try to remove a directory which has a hidden file in it (other
than .Acorn, which is already handled explicitly) the remove will
fail.