Protocol: support EC_FS_FUNC_CAT_HEADER.
This is a bit iffy and almost untested, since I wasn't able to find
any real client code that knew what to expect from this request. I
did find an implementation of it, though, in the BBC BASIC Econet
Level 1 file server that came with BeebEm; so I ran that, looked at
what it returned, constructed an implementation that matched it, and
checked with a network monitor that the two implementations returned
packets with matching formats.

This is also a second piece of code (after EC_FS_GET_INFO_DIR) which
needs to know if a directory is Owner or Public. I haven't bothered
putting in an abstraction both can call, but I have at least put in
comments cross-referencing each to the other so that anyone who
improves that area of aund won't forget :-)

Having done this request, aund now supports every file server
request type listed in the Econet System User Guide. Hooray!